refactor skymap to healpix_dataset
