+++ src/lsdb/catalog/catalog.py
@@ -414,40 +414,3 @@ class Catalog(HealpixDataset):
         )
         hc_catalog = hc.catalog.Catalog(new_catalog_info, alignment.pixel_tree)
         return Catalog(ddf, ddf_map, hc_catalog)
-
-    def skymap_data(
-        self, func: Callable[[pd.DataFrame, HealpixPixel], Any], **kwargs
-    ) -> Dict[HealpixPixel, Delayed]:
-        """Perform a function on each partition of the catalog, returning a dict of values for each pixel.
-
-        Args:
-            func (Callable[[pd.DataFrame, HealpixPixel], Any]): A function that takes a pandas
-                DataFrame with the data in a partition, the HealpixPixel of the partition, and any other
-                keyword arguments and returns an aggregated value
-            **kwargs: Arguments to pass to the function
-
-        Returns:
-            A dict of Delayed values, one for the function applied to each partition of the catalog
-        """
-
-        partitions = self.to_delayed()
-        results = {
-            pixel: delayed(func)(partitions[index], pixel, **kwargs)
-            for pixel, index in self._ddf_pixel_map.items()
-        }
-        return results
-
-    def skymap(self, func: Callable[[pd.DataFrame, HealpixPixel], Any], **kwargs):
-        """Plot a skymap of an aggregate function applied over each partition with a Mollweide projection.
-
-        Args:
-            func (Callable[[pd.DataFrame], HealpixPixel, Any]): A function that takes a pandas DataFrame and
-                the HealpixPixel the partition is from and returns a value
-            **kwargs: Arguments to pass to healpy.mollview function
-        """
-
-        smdata = self.skymap_data(func, **kwargs)
-        pixels = list(smdata.keys())
-        results = dask.compute(*[smdata[pixel] for pixel in pixels])
-        result_dict = {pixels[i]: results[i] for i in range(len(pixels))}
-        plot_skymap(result_dict)
